Refactor to move boundary related implementation into a common place.
